Redux SiteProvider in progress, needs refresh replacement in components
