+++ front/src/services/site/actions.ts
@@ -4,6 +4,8 @@ import * as types from './types';
 import { SitesPageQuery } from 'services/site/model/SitesPageQuery';
 import { AdminViewsProviderQuery } from 'services/site/model/AdminViewsProviderQuery';
 import { CreateSiteInput, UpdateSiteInput } from 'services/site/model/InputTypes';
+import { SiteFragment as SiteProviderQuery } from 'services/site/model/SiteFragment';
+
 
 export const fetchAdminUserSite = () : types.SiteActionTypes => ({
     type: types.FETCH_ADMIN_SITE_VIEW_SEND
@@ -31,6 +33,23 @@ export const fetchSitesPageSuccess = (payload: SitesPageQuery) : types.SiteActio
 
 export const fetchSitesPageError = (message: string) : types.SiteActionTypes => ({
     type: types.FETCH_SITES_PAGE_ERROR,
+    payload: { message }
+});
+
+
+export const fetchSiteProvider = (id: number, url: string) : types.SiteActionTypes => ({
+    type: types.FETCH_SITE_PROVIDER_SEND,
+    id,
+    url
+});
+
+export const fetchSiteProviderSuccess = (payload: SiteProviderQuery) : types.SiteActionTypes => ({
+    type: types.FETCH_SITE_PROVIDER_SUCCESS,
+    payload
+});
+
+export const fetchSiteProviderError = (message: string) : types.SiteActionTypes => ({
+    type: types.FETCH_SITE_PROVIDER_ERROR,
     payload: { message }
 });
