+++ front/src/services/site/api.ts
@@ -15,6 +15,11 @@ export const fetchSitesPage = () => {
     return callGraphql(ENDPOINT, query.SITES_PAGE_QUERY, {});
 }
 
+export const fetchSiteProvider = (id, url) => {
+    return callGraphql(ENDPOINT, query.SITE_PROVIDER_QUERY,
+        { input: id, url});
+}
+
 export const deleteSite = ( id ) => {
     return callGraphql(ENDPOINT, mutate.DELETE_SITE_MUTATION, 
         { input: { id} });
