+++ front/src/services/site/mutations.ts
@@ -1,5 +1,5 @@
 
-import { SITE_FRAGMENT } from 'services/site/model/SiteFragment'; //TODO Move/delete site fragment from SiteProvider.tsx when we do redux there.
+import { SITE_FRAGMENT } from 'services/site/SiteFragments'; //TODO Move/delete site fragment from SiteProvider.tsx when we do redux there.
 
 
 export const DELETE_SITE_MUTATION = `
