+++ front/src/services/site/sagas.ts
@@ -39,6 +39,26 @@ function* getSitesPage(action) {
     }
 }
 
+function* getSiteProvider(action) {
+            console.log("SAGA get Site Provider", action);
+
+    try {
+        let response = yield call(() => api.fetchSiteProvider(action.id, action.url));
+        if(response) {
+            yield put(actions.fetchSiteProviderSuccess(response.data));
+            return response;
+        }
+        else {
+            yield put(actions.fetchSiteProviderError(response.message));
+        }
+    }
+    catch(err) {
+        console.log(err);
+        yield put(actions.fetchSiteProviderError(err.message));
+    }
+}
+
+
 function* deleteSite(action) { 
     const currentSites = yield select(getCurrentSites)
     try {
@@ -105,6 +125,7 @@ function* updateSite(action) {
 export default function* userSagas() {
     yield takeLatest(types.FETCH_ADMIN_SITE_VIEW_SEND, getAdminSiteView);
     yield takeLatest(types.FETCH_SITES_PAGE_SEND, getSitesPage);
+    yield takeLatest(types.FETCH_SITE_PROVIDER_SEND, getSiteProvider);
     yield takeLatest(types.DELETE_SITE_SEND, deleteSite);
     yield takeLatest(types.CREATE_SITE_SEND, createSite);
     yield takeLatest(types.UPDATE_SITE_SEND, updateSite);
