+++ front/src/services/site/types.ts
@@ -1,6 +1,8 @@
+import { SiteFragment as SiteProviderQuery } from 'services/site/model/SiteFragment';
 import { CreateSiteInput, UpdateSiteInput } from 'services/site/model/InputTypes';
 import { SitesPageQuery } from 'services/site/model/SitesPageQuery';
 import { AdminViewsProviderQuery } from 'services/site/model/AdminViewsProviderQuery';
+
 
 export const FETCH_ADMIN_SITE_VIEW_SEND = 'FETCH_ADMIN_SITE_VIEW_SEND';
 export const FETCH_ADMIN_SITE_VIEW_SUCCESS = 'FETCH_ADMIN_SITE_VIEW_SUCCESS';
@@ -9,6 +11,10 @@ export const FETCH_ADMIN_SITE_VIEW_ERROR = 'FETCH_ADMIN_SITE_VIEW_ERROR';
 export const FETCH_SITES_PAGE_SEND = 'FETCH_SITES_PAGE_SEND';
 export const FETCH_SITES_PAGE_SUCCESS = 'FETCH_SITES_PAGE_SUCCESS';
 export const FETCH_SITES_PAGE_ERROR = 'FETCH_SITES_PAGE_ERROR';
+
+export const FETCH_SITE_PROVIDER_SEND = 'FETCH_SITE_PROVIDER_SEND';
+export const FETCH_SITE_PROVIDER_SUCCESS = 'FETCH_SITE_PROVIDER_SUCCESS';
+export const FETCH_SITE_PROVIDER_ERROR = 'FETCH_SITE_PROVIDER_ERROR';
 
 export const DELETE_SITE_SEND = 'DELETE_SITE_SEND';
 export const DELETE_SITE_SUCCESS = 'DELETE_SITE_SUCCESS';
@@ -24,12 +30,14 @@ export const UPDATE_SITE_ERROR = 'UPDATE_SITE_ERROR';
 
 export interface SiteState {
     isFetchingAdminSiteView: boolean,
-    adminSiteView: AdminViewsProviderQuery | undefined;
+    adminSiteView: AdminViewsProviderQuery | undefined,
     isFetchingSitesPage: boolean,
     sitesData: any | SitesPageQuery | undefined,
     isDeletingSite: boolean,
     isCreatingSite: boolean,
     isUpdatingSite: boolean,
+    isFetchingSiteProvider: boolean,
+    siteProvider: SiteProviderQuery | undefined,
 }
 
 export interface SiteDataError {
@@ -63,6 +71,23 @@ export interface FetchSitesPageErrorAction {
     type: typeof FETCH_SITES_PAGE_ERROR,
     payload: SiteDataError
 };
+
+export interface FetchSiteProviderSendAction {
+    type: typeof FETCH_SITE_PROVIDER_SEND,
+    id: number,
+    url: string
+};
+
+export interface FetchSiteProviderSuccessAction {
+    type: typeof FETCH_SITE_PROVIDER_SUCCESS,
+    payload: SiteProviderQuery
+};
+
+export interface FetchSiteProviderErrorAction {
+    type: typeof FETCH_SITE_PROVIDER_ERROR,
+    payload: SiteDataError
+};
+
 
 export interface DeleteSiteSendAction {
     type: typeof DELETE_SITE_SEND,
@@ -114,4 +139,5 @@ export interface UpdateSiteErrorAction {
 
 export type SiteActionTypes = FetchAdminSiteViewSendAction | FetchAdminSiteViewSuccessAction | FetchAdminSiteViewErrorAction |
     FetchSitesPageSendAction | FetchSitesPageSuccessAction | FetchSitesPageErrorAction | DeleteSiteSendAction | DeleteSiteSuccessAction |
-    DeleteSiteErrorAction  | CreateSiteSendAction | CreateSiteSuccessAction | CreateSiteErrorAction  | UpdateSiteSendAction | UpdateSiteSuccessAction | UpdateSiteErrorAction  ;
+    DeleteSiteErrorAction  | CreateSiteSendAction | CreateSiteSuccessAction | CreateSiteErrorAction  
+    | UpdateSiteSendAction | UpdateSiteSuccessAction | UpdateSiteErrorAction | FetchSiteProviderSendAction | FetchSiteProviderSuccessAction | FetchSiteProviderErrorAction ;
